crypto04/src/src/requirements.txt
Flask==2.2.4
gunicorn==20.1.0
pycryptodome==3.19.0
flask_session==0.5.0